// src/app/shared/models/Package/paclage-return-list-model.ts
export class PackageReturnListModel {
  id: number;
  ssName: string;
  shipWeight: number;
  countryCode: string;
  countryCnName: string;
  countryEnName: string;
  latestScanTime: Date;
  latestScanWeight: number;
  channelName: string;
  spPackageId: string;
  trackingNumber: string;
  amount: number;
  shippedTime: Date;
  shippedDays: number;
  currencyCode: string;
  packageNumber: string;
  packageId: number;
  warehouseId: number;
  isMarkShipped: boolean;
}